Cluster Random Sample Size

Gives the number of clusters (pairs of clusters for paired designs) required for cluster randomized trials

cluster_random_sample_size(alpha, power, coef_var = 0.25, A = 2, pi, pc, CS)

Arguments

alpha

significance level

power

the desired power

coef_var

between-cluster variation, use 0.25 if unknown as this is the estimated maximum

A

Adjustment for loss of degrees of freedom (1 for unpaired trials, 2 in paired trials)

pi

proportion in the intervention group

pc

proportion in the control group

CS

an estimate of the cluster size

Value

number of clusters required in each treatment group in an unmatched design, and total pairs in a paired design.

References

Hayes, R. J., & Moulton, L. H. (2017). Cluster randomised trials. Chapman and Hall/CRC.

Examples

# The matched trial protocol required 80% power of detecting a 50% reduction of annual HIV incidence from an assumed 1% in the control arm (π0 = 0.02 over 2 years) to 0.5% in the intervention arm (π1 = 0.01). Thus, with m = 1000 adults followed up in each cluster, the number of clusters required per treatment arm in a matched design was gives 6.8 = 7 cluster pairs
cluster_random_sample_size(alpha=0.05, power=0.8, coef_var=0.25, A =2, pi=0.01, pc=0.02, CS=1000)
#> [1] 7